refactor: reorganizes components folder to provide a better separation for each type of component
