+++ src/lib/components/atoms/index.ts
@@ -1,0 +1,14 @@
+/**
+ * Atoms
+These are the most basic, indivisible components — like the HTML tags of your design system. They include:
+
+- Single buttons
+
+- Inputs
+
+- Labels
+
+- Icons
+
+- Typography elements (e.g., headings)
+ */
