+++ src/lib/components/atoms/sectionTitle/SectionTitleProps.ts
