+++ src/lib/components/atoms/sectionTitle/sectionTitle.svelte
