+++ src/lib/components/atoms/skill/SkillProps.ts
@@ -1,0 +1,3 @@
+import type { Skill } from '../../../../apis/dtos/PortfolioData';
+
+export type SkillProps = Skill;
