--- src/lib/components/experience/ExperienceProps.ts
@@ -1,5 +1,0 @@
-import type { Experience } from '../../../apis/dtos/PortfolioData';
-
-export type ExperienceProps = {
-	experience: Experience;
-};
