--- src/lib/components/footer/FooterProps.ts
@@ -1,5 +1,0 @@
-import type { Info } from '../../../apis/dtos/PortfolioData';
-
-export type FooterProps = {
-	info: Info;
-};
