+++ src/lib/components/molecules/experience/ExperienceProps.ts
@@ -1,0 +1,5 @@
+import type { Experience } from '../../../../apis/dtos/PortfolioData';
+
+export type ExperienceProps = {
+	experience: Experience;
+};
