+++ src/lib/components/molecules/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Molecules
+These are groups of atoms working together as a single unit. They’re still relatively simple and often form reusable form elements or interactive units.
+
+Example:
+
+- Input + Label
+
+- Button + Icon
+
+- Avatar + Username
+ */
