+++ src/lib/components/molecules/project/ProjectProps.ts
@@ -1,0 +1,3 @@
+import type { Project } from '../../../../apis/dtos/PortfolioData';
+
+export type ProjectProps = Project;
