--- src/lib/components/navBar/NavBarProps.ts
@@ -1,6 +1,0 @@
-import type { Info, Project } from '../../../apis/dtos/PortfolioData';
-
-export type NavBarProps = {
-	info: Info;
-	projects: Project[];
-};
