+++ src/lib/components/organisms/footer/FooterProps.ts
@@ -1,0 +1,5 @@
+import type { Info } from '../../../../apis/dtos/PortfolioData';
+
+export type FooterProps = {
+	info: Info;
+};
