+++ src/lib/components/organisms/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Organisms
+These are more complex UI sections formed by combining molecules and atoms. Organisms usually represent a distinct section of a page or feature.
+
+Examples:
+
+- Navigation bar
+
+- Login form
+
+- Product card
+ */
