+++ src/lib/components/organisms/navBar/NavBarProps.ts
@@ -1,0 +1,6 @@
+import type { Info, Project } from '../../../../apis/dtos/PortfolioData';
+
+export type NavBarProps = {
+	info: Info;
+	projects: Project[];
+};
