--- src/lib/components/project/ProjectProps.ts
@@ -1,3 +1,0 @@
-import type { Project } from '../../../apis/dtos/PortfolioData';
-
-export type ProjectProps = Project;
