--- src/lib/components/skill/SkillProps.ts
@@ -1,3 +1,0 @@
-import type { Skill } from '../../../apis/dtos/PortfolioData';
-
-export type SkillProps = Skill;
