+++ src/lib/components/templates/about-me/AboutMeProps.ts
@@ -1,0 +1,5 @@
+import type { Info } from '../../../../apis/dtos/PortfolioData';
+
+export type AboutMeSectionProps = {
+	info: Info;
+};
