+++ src/lib/components/templates/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Templates
+Templates are page-level layouts composed of organisms placed in a structure. They define the layout without real content.
+
+Examples:
+
+- Dashboard layout
+
+- Blog post layout
+ */
