+++ src/lib/components/templates/main/MainSectionProps.ts
@@ -1,0 +1,5 @@
+import type { Info } from '../../../../apis/dtos/PortfolioData';
+
+export type MainSectionProps = {
+	info: Info;
+};
