+++ src/lib/components/templates/projects/ProjectsProps.ts
@@ -1,0 +1,5 @@
+import type { Project } from '../../../../apis/dtos/PortfolioData';
+
+export type ProjectsProps = {
+	projects: Project[];
+};
