--- src/lib/sections/about-me/AboutMeProps.ts
@@ -1,5 +1,0 @@
-import type { Info } from '../../../apis/dtos/PortfolioData';
-
-export type AboutMeSectionProps = {
-	info: Info;
-};
