--- src/lib/sections/main/MainSectionProps.ts
@@ -1,5 +1,0 @@
-import type { Info } from '../../../apis/dtos/PortfolioData';
-
-export type MainSectionProps = {
-	info: Info;
-};
