--- src/lib/sections/projects/ProjectsProps.ts
@@ -1,5 +1,0 @@
-import type { Project } from '../../../apis/dtos/PortfolioData';
-
-export type ProjectsProps = {
-	projects: Project[];
-};
